Refactor tests - split into unit and feature Unit tests can be run by Github Actions
